refactor(chat): begin chat_service.dart decomposition — extract group read accessors
chat_service.dart is the project's main god file (>10k lines, far over the
500-line cap). This starts breaking it apart using the proven part-file pattern
(same one used for character_gen_service): cohesive method groups move into
`part of` files that share the library's imports and private members, so the move
is verbatim with zero behaviour change and no call-site edits.

First extraction: the per-group-character realism/needs READ accessors consumed
by the sidebar / member cards (getEmotionForGroupCharacter, getNeedsForGroupCharacter,
getAffectionForGroupCharacter, getTrustForGroupCharacter, getArousalForGroupCharacter,
getFixationForGroupCharacter, getEmotionIntensityForGroupCharacter,
getFixationLifespanForGroupCharacter, getTopUrgentNeedsForGroupCharacter,
getRealismStateForGroupCharacter, resetRealismForGroupCharacter) → new
lib/services/chat/chat_service_group_read.dart as `extension ChatServiceGroupRead
on ChatService`. They were pure reads over the private `_groupRealism` map plus
one reset helper, with no orchestration, so they are a clean, self-contained unit.

chat_service.dart drops from 10,506 to 10,374 lines; the part-file infrastructure
is now in place for further extractions. flutter analyze clean; full non-golden
test suite (1334 tests) green.
